Make serializeVarintCommon SCARY
Move the part that doesn't depend on template arugment to a
serializeVarintCommonImpl function. Also it makes sense to make it
out of line.

Use the impl function but not the template one when serializing
the size of length delimited fields. We trust out selves that this
type is compatible with serializeVarintCommon.

Pick-to: 6.8
Task-number: QTBUG-128812
Change-Id: Ifcda7ebf7b2dee209c258d64724908022cf66de9
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit e2ef159b2e85f4cfc99e36ffb5c956f8bfc066d2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>