Add the map name clashing syntax test
Protoc disallows using the <field name>Entry if the respective
<field name> map field is defined. But meanwhile it allows using
<field name>_Entry as the field name, which ends up as
<field name>Entry in the generated code. The difference is in the
letter case is used for the alias we add for map entries and the
actual field name.

As the conclusion - test passes, but we should keep it in our test
suite to ensure we compliant with the protoc rules.

Pick-to: 6.8
Change-Id: I8a8d6796b6c5564b211bd11fdd23102b0031031e
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit c42df9cad0a192f241f5e9c52462b8b30fff70dd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>